Moved checkout services into a specific folder under app/services
